requirements.txt
altair==4.1.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
black==20.8b1
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cloudscraper==1.2.56
colorama==0.4.4
decorator==4.4.2
defusedxml==0.7.1
docutils==0.16
entrypoints==0.3
Flask==1.1.2
gitdb==4.0.7
GitPython==3.1.14
idna==2.10
importlib-metadata==3.7.3
ipykernel==5.5.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==23.0.1
loguru==0.5.3
MarkupSafe==1.1.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.1
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pkginfo==1.7.0
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==3.15.6
ptyprocess==0.7.0
pyarrow==3.0.0
pycparser==2.20
pydeck==0.6.1
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-box==5.3.0
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
readme-renderer==29.0
regex==2021.3.17
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
Send2Trash==1.5.0
six==1.15.0
smmap==4.0.0
streamlit==0.79.0
terminado==0.9.3
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
twine==3.4.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.1
